// dist/api/user/types.ts
/* eslint-disable */
// tslint:disable
/**
 * MicroService Demo
 * 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface UserApi {
	listUser: () => Promise<ListUserResponse>
	postUser: (request: Api.User) => Promise<PostUserResponse>
	deleteUser: (userId: string) => Promise<DeleteUserResponse>
	getUser: (userId: string) => Promise<GetUserResponse>
	updateUser: (userId: string, request: Api.User) => Promise<UpdateUserResponse>
}

export type ListUserResponse = ListUser200Response

export interface ListUser200Response {
	status: 200
	body: Api.User[]
	headers?: never
}

export type PostUserResponse = PostUser201Response | PostUser400Response | PostUser409Response

export interface PostUser201Response {
	status: 201
	body: string
	headers?: never
}

export interface PostUser400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export interface PostUser409Response {
	status: 409
	body: Api.ApiResponse
	headers?: never
}

export type DeleteUserResponse = DeleteUser200Response | DeleteUser400Response | DeleteUser404Response

export interface DeleteUser200Response {
	status: 200
	body: Api.ApiResponse
	headers?: never
}

export interface DeleteUser400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export interface DeleteUser404Response {
	status: 404
	body: Api.ApiResponse
	headers?: never
}

export type GetUserResponse = GetUser200Response | GetUser400Response | GetUser404Response

export interface GetUser200Response {
	status: 200
	body: Api.User
	headers?: never
}

export interface GetUser400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export interface GetUser404Response {
	status: 404
	body: Api.ApiResponse
	headers?: never
}

export type UpdateUserResponse = UpdateUser200Response | UpdateUser400Response | UpdateUser404Response

export interface UpdateUser200Response {
	status: 200
	body: Api.ApiResponse
	headers?: never
}

export interface UpdateUser400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export interface UpdateUser404Response {
	status: 404
	body: Api.ApiResponse
	headers?: never
}